feat: adds the ERI_CELL_ANGLES keyword
This enables the specification of custom cell angles used during the ERI
GPW calculation within the active space module. A key use-case for this
is to specify a cubic super cell to be used with the Martyna-Tuckerman
Poisson solver even though (for example) a non-cubic cell is used during
the outer SCF calculation.